pyproject.toml
[project]
name = "codesight"
version = "0.2.0"
description = "AI-powered document search engine — hybrid BM25 + vector retrieval with Claude answer synthesis"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers>=3.0",
    "lancedb>=0.17",
    "pandas>=2.0",
    "pathspec>=0.12",
    "pydantic>=2.0",
    "anthropic>=0.40",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "python-pptx>=0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]
demo = [
    "streamlit>=1.40",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codesight"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
